+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Vector distance squared static method.
+- Circle from vector static constructor method.
 
 ## [5.2.0] - 2025-01-20